save role

+++ database/migrations/2023_07_05_024623_create_tasks_table.php
@@ -15,10 +15,11 @@ class CreateTasksTable extends Migration
     {
         Schema::create('tasks', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('title');
             $table->boolean('is_complete');
             $table->timestamps();
+
             $table->foreign('user_id')->references('id')->on('users');
         });
     }
